requirements.txt
pydicom
Pillow
